backend/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.7.0
django-extensions==3.2.3
django-ratelimit==4.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
packaging==25.0
pillow==10.4.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
pytz==2025.2
sqlparse==0.5.1
tzdata==2024.2
